+++ NothingWebApi/Services/NothingService.cs
@@ -10,13 +10,16 @@ namespace NothingWebApi.Services;
 /// Сервис бизнес логики
 /// </summary>
 /// <param name="dbContext">Контекст базы данных</param>
+/// <param name="logger">Экземпляр <see cref="ILogger"/></param>
 /// <param name="mapper"><see cref="Mapper"/></param>
 public class NothingService(
     NothingWebApiDbContext dbContext,
+    ILogger<NothingService> logger,
     IMapper mapper)
     : INothingService
 {
     private readonly NothingWebApiDbContext _dbContext = dbContext;
+    private readonly ILogger<NothingService> _logger = logger;
     private readonly IMapper _mapper = mapper;
 
     /// <summary>
@@ -24,11 +27,21 @@ public class NothingService(
     /// </summary>
     /// <param name="cancellationToken">Токен отмены</param>
     /// <returns>Список моделей</returns>
-    public Task<NothingModelDto[]> GetAsync(CancellationToken cancellationToken = default)
+    public Task<NothingModelDto[]> Get(CancellationToken cancellationToken = default)
     {
-        return _dbContext.NothingModels.AsNoTracking()
-            .Select(model => _mapper.Map<NothingModelDto>(model))
-            .ToArrayAsync(cancellationToken);
+        try
+        {
+            return _dbContext.NothingModels
+                .AsNoTracking()
+                .OrderBy(model => model.Id)
+                .Select(model => _mapper.Map<NothingModelDto>(model))
+                .ToArrayAsync(cancellationToken);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex.ToString());
+            throw;
+        }
     }
 
     /// <summary>
@@ -37,12 +50,24 @@ public class NothingService(
     /// <param name="id">Идентификатор модели</param>
     /// <param name="cancellationToken">Токен отмены</param>
     /// <returns>Объект модели</returns>
-    public Task<NothingModelDto> GetAsync(int id, CancellationToken cancellationToken = default)
+    public async Task<NothingModelDto> Get(int id, CancellationToken cancellationToken = default)
     {
-        return _dbContext.NothingModels.AsNoTracking()
-            .Where(model => model.Id == id)
-            .Select(model => _mapper.Map<NothingModelDto>(model))
-            .SingleAsync(cancellationToken);
+        try
+        {
+            var nothingModelDto = await _dbContext.NothingModels
+                .AsNoTracking()
+                .Where(model => model.Id == id)
+                .Select(model => _mapper.Map<NothingModelDto>(model))
+                .SingleOrDefaultAsync(cancellationToken);
+            if(nothingModelDto == null)
+                throw new ArgumentException($"Не удалось найти модель с идентификатором {id}.");
+            return nothingModelDto;
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex.ToString());
+            throw;
+        }
     }
 
     /// <summary>
@@ -54,16 +79,24 @@ public class NothingService(
     /// <exception cref="ArgumentNullException">
     /// Ошибка валидации входных данных
     /// </exception>
-    public async Task<NothingModelDto> CreateAsync(
+    public async Task<NothingModelDto> Create(
         CreateNothingModelDto createNothingModelDto,
         CancellationToken cancellationToken = default)
     {
-        if(string.IsNullOrEmpty(createNothingModelDto.Name))
-            throw new ArgumentNullException(nameof(createNothingModelDto.Name), "Name cannot be null or empty.");
-        var model = _mapper.Map<NothingModel>(createNothingModelDto);
-        await _dbContext.NothingModels.AddAsync(model, cancellationToken);
-        await _dbContext.SaveChangesAsync(cancellationToken);
-        return _mapper.Map<NothingModelDto>(model);
+        try
+        {
+            if(string.IsNullOrEmpty(createNothingModelDto.Name.Trim()))
+                throw new ArgumentNullException(nameof(createNothingModelDto.Name), "Имя не может быть пустым.");
+            var model = _mapper.Map<NothingModel>(createNothingModelDto);
+            await _dbContext.NothingModels.AddAsync(model, cancellationToken);
+            await _dbContext.SaveChangesAsync(cancellationToken);
+            return _mapper.Map<NothingModelDto>(model);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex.ToString());
+            throw;
+        }
     }
 
     /// <summary>
@@ -75,17 +108,27 @@ public class NothingService(
     /// <exception cref="ArgumentNullException">
     /// Ошибка валидации входных данных
     /// </exception>
-    public async Task<NothingModelDto> UpdateAsync(
+    public async Task<NothingModelDto> Update(
         UpdateNothingModelDto updateNothingModelDto,
         CancellationToken cancellationToken = default)
     {
-        if (string.IsNullOrEmpty(updateNothingModelDto.Name))
-            throw new ArgumentNullException(nameof(updateNothingModelDto.Name), "Name cannot be null or empty.");
-        var model = await _dbContext.NothingModels
-            .SingleAsync(model => model.Id == updateNothingModelDto.Id, cancellationToken);
-        model.Name = updateNothingModelDto.Name;
-        await _dbContext.SaveChangesAsync(cancellationToken);
-        return _mapper.Map<NothingModelDto>(model);
+        try
+        {
+            if (string.IsNullOrEmpty(updateNothingModelDto.Name.Trim()))
+                throw new ArgumentNullException(nameof(updateNothingModelDto.Name), "Имя не может быть пустым.");
+            var model = await _dbContext.NothingModels
+                .SingleOrDefaultAsync(model => model.Id == updateNothingModelDto.Id, cancellationToken);
+            if(model == null)
+                throw new ArgumentException($"Не удалось найти модель с идентификатором {updateNothingModelDto.Id}.");
+            model.Name = updateNothingModelDto.Name.Trim();
+            await _dbContext.SaveChangesAsync(cancellationToken);
+            return _mapper.Map<NothingModelDto>(model);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex.ToString());
+            throw;
+        }
     }
 
     /// <summary>
@@ -94,12 +137,22 @@ public class NothingService(
     /// <param name="id">Идентификатор модели</param>
     /// <param name="cancellationToken">Токен отмены</param>
     /// <returns>Объект модели</returns>
-    public async Task<NothingModelDto> DeleteAsync(int id, CancellationToken cancellationToken = default)
+    public async Task<NothingModelDto> Delete(int id, CancellationToken cancellationToken = default)
     {
-        var model = await _dbContext.NothingModels
-            .SingleAsync(model => model.Id == id, cancellationToken);
-        _dbContext.NothingModels.Remove(model);
-        await _dbContext.SaveChangesAsync(cancellationToken);
-        return _mapper.Map<NothingModelDto>(model);
+        try
+        {
+            var model = await _dbContext.NothingModels
+                .SingleOrDefaultAsync(model => model.Id == id, cancellationToken);
+            if(model == null)
+                throw new ArgumentException($"Не удалось найти модель с идентификатором {id}.");
+            _dbContext.NothingModels.Remove(model);
+            await _dbContext.SaveChangesAsync(cancellationToken);
+            return _mapper.Map<NothingModelDto>(model);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex.ToString());
+            throw;
+        }
     }
 }
